udn, cni, tests: Ensure we report the UDN interface via CNI result
Whenever we invoke `cmdAddWithGetCNIResultFunc` it will return a result
with 2 interfaces: the first for the host side veth, and the second for
the pod side veth (featuring the sandbox attribute). Check the source
code at [0] to understand how it happens.

Now that we have changed the production code to *also* reply with the
result of adding the UDN interface, we are in fact returning 4
interfaces on the result of `cmdAddWithGetCNIResultFunc`:

index: interface
----------------
0: host side primary UDN
1: pod side primary UDN
2: host side cluster default network
3: pod side cluster default network

[0] -
https://github.com/ovn-org/ovn-kubernetes/blob/a1e88dcd0b79b6c5eb033daea178708993f8664e/go-controller/pkg/cni/helper_linux.go#L650

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>
Co-authored-by: Ram Lavi <ralavi@redhat.com>