udn,template: Enable NAD rendering for CUDN objects
Since the NAD metadata and the CNI config args 'name' & 'netAttachDefName'
depends on the given object kind and metadata, its imposable to reuse for the
cluster UDN CRD objects.

Refactor rendering code to enable reuse for for UDN and CUDN object:
- Decouple the NAD spec rendering and the NAD metadata and & CNI config args.
- Generlize factory to accept UDN or CUDN objects, and the CNI conf NAD name.
  And generate the NAD network-name, and owner-referance accordingly.

This is preliminary change in order to resue code for the CUDN controller
in follow-up commit.

Signed-off-by: Or Mergi <ormergi@redhat.com>