fixes overzealous deletion of SNAT in egressIP
currently when creating egressIP in a dualstack cluster regardless of if
the egressip is ipv4 or ipv6 when disable-snat-multiple-gws is set the
code removes both ipv4 and ipv6 snats from the database. This means that
the pod will not be able to communicate with the cluster correctly
becuase the Gateway Router is missing an SNAT and the traffic will be
dropped.

Additionally the testing looks a little different from the others
because we do not correctly setup dualstack clusters when running unit
tests

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>