Fix missing ICMP SNAT for L2 UDNs GR
On GRs we have an SNAT from the join port IP address to the external IP
address for ICMP needs frag sent by OVN to the external network.

In L2 UDNs, even though there is no join network, the internal port has
two IPs, one for the pod subnet and another for the join subnet. The
ICMP source address will be the first one in lexicographical order which
might be the pod subnet one.

The proper SNAT for the GR pod subnet address gets created in the gateway
initialization. But:
1 when DisableSNATMultipleGWs=false, it is redundant because there is
  already a SNAT for the whole pod subnet
2 when DisableSNATMultipleGWs=true we end up with no SNAT because
  cleanupStalePodSNATs cleans it up becaause it never expected that
  SNAT

This commit fixes #2

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>