Split initGateway()
Split initGateway into two parts: initGatewayPreStart and initGatewayMainStart.

A later commit ("Split gateway start()") will change the execution flow so as to run initGatewayPreStart /before/ the Secondary Node Network Controller (SNNC) is started; initGatewayMainStart will run with the rest of Default Node Network Controller (DNNC) initialization.

The motivation behind this is that SNNC needs to reference openflow manager created in DNNC and SNNC is started before the rest of DNNC functionality is started.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>