Skip handling NetPol events on non primary networks
When primary UDN exists for the namespace, the current implementation configured
network policy for both UDN and default network. The default network traffic
has its own ACLs that deny almost everything already so handling network policy
for default is unnecessary and not an optimal solution as it programs another
set of port groups, acls and address sets which are never going to be used.
Hence this commit skips handling network policy events on the base network
controller when there is a mismatch on active primary network.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>