e2e flake fix: kubevirt
This PR addresses a flake in KubeVirt E2E tests, improving
the connectivity tests.

It also introduces the use of an environment check for
KUBEVIRT_SKIP_MIGRATE_POST_COPY to avoid failing the
kubevirtv1.MigrationPostCopy mode, which is currently
used in the GitHub workflow.

Lastly, it makes it possible to skip tests in deployments
where OVS is installed on the node instead of running as
a pod.

Fixes: https://github.com/ovn-org/ovn-kubernetes/issues/4709
Signed-off-by: Flavio Fernandes <ffernandes@nvidia.com>