interconenct: Ask ovn-northd to bind the remote ports.
Presently for each remote node, ovnkube after creating the
remote port in the OVN Northbound db transit switch, binds the port to
the remote chassis in the Southbound database.  In order to do this,
we have to wait for the remote logical ports (for each of the
remote zone nodes) to appear in the OVN Southbound database.
This can lead to race conditions.  Instead ask ovn-northd to
bind to the remote chassis.  OVN after this commit [1] already
supports this.  We just need to set the requested-chassis
option in the logical port.

[1] - https://github.com/ovn-org/ovn/commit/170d3e5f19f0fcd760db068cda29fc1b7f1d7ef5

Jira: https://issues.redhat.com/browse/OCPBUGS-42616

Signed-off-by: Numan Siddique <numans@ovn.org>