Fix UTs for UDNs: don't set exclude_ips
There is an unrelated UT failure that I hit
in this PR because of timing issues. This commit
is in no way related to my original PR's issue but
I need to satisfy CI.

A few general issues for which I added FIXME's in code:

1) Why do we set exclude_ips on switch at startup and then
using UpdateNodeSwitchExcludeIPs we seem to remove them
from the switch when we add the management port IPs. We
don't use OVN IPAM. OVNK seems to be providing the IPs
for the LSPs always.
2) UT for primary should not have exclude_ips set since
we remove them using UpdateNodeSwitchExcludeIPs function
from syncManagementPort. This commit fixes this
3) I don't know why the UT expects the excludes_ips to be
present on the switch for secondary networks when secondary
networks don't even have managementports and don't use .2IP.
But that problem is not related to this PR.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>