ANP: Use configured primary network pod ips and logical port name
The admin policy controller currently uses the default network to retrieve pod IPs
and its logical port name. But this is not true when pod is configured with
user defined network as its primary network. So this commit updates the controller
to use appropriate netinfo object to retrieve those parameters. This way ANP
can be supported for UDN though the ANP controller is initialized from the default
network controller.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>