Disable kube manager service-lb-controller for kind
It seems that v1.31.1 introduced a bug in kube manager's
service-lb-controller. Since we don't use a cloud provider, the
controller is not fully initialized and started. However, its handlers
are added to the informer and they do run. And when they do, it crashes
because it is not fully initialized.

Probably introduced through:
https://github.com/kubernetes/kubernetes/commit/50c12437604b0cd5a73514389409fc2fde8b91bd

Disable service-lb-controller since it is not used anyway.

bootstrap-signer-controller and token-cleaner-controller need to be
added since they are not default and would otherwise be added by kind
but not if we are overriding.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>