node_annotations: Make GetNodeHostAddrs() return stable results.
The host address annotations were parsed as a set of strings and sets
are unordered.  In consequence GetNodeHostAddrs() might return
differently ordered host addresses when invoked on the same annotation.

In some cases this can cause DeepEqual() comparisons of objects that
contain the parsed annotation results to fail even though the annotation
itself didn't change.

One such example is with NodeTracker.updateNodeInfo() which compares two
nodeInfo structures that contain parsed lists of host addresses.  The
unstable parsing was causing unexpected full service resyncs when
nothing relevant had actually changed on the node annotation.

This becomes more visible when testing with a larger number of UDNs
because the NodeTracker objects are created per UDN.

Without this fix, on a 20 node kind cluster when creating 100 L3 user
defined networks with one pod each, pprof CPU traces show
Controller.syncService() -> Controller.configureUDNEnabledServiceRoute()
taking up ~35% of the CPU time.  Some of the services were resynced
thousands of times on some of the UDNs.

With the fix, no service resync happens on already configured UDNs in
this case.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>