Ensure UDN namespace is configured when MutiNetPolicy is disabled
The namespace is currently not getting configured in ovnk and ovn for user
defined network when multi network policy is disabled for the cluster. But
multinet policy is not a requirement for UDN network policies, hence fixing
it so that namespaces get configured for those namespaces on secondary
controller and network policies can be created for the same.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>