Egress node: add default route to external also in update path
The "add" code for egress nodes eventually adds the default route on ovn_cluster_router to the gw router (CreateDefaultRouteToExternal), while the update path does not.

If ever the "add" event on an egress node failed, it'll be retried. Any update event on the node happening /before/ the retry add is done will result in an update even being processed first. In this case, the egress node will be added through the update path and the default route won't be added.

Let's add the default route also in the update path.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>