Adds startup probe to UDN e2e
It was seen in downstream that this container can take longer than a
second for its http server to start sometimes:

https://prow.ci.openshift.org/view/gs/test-platform-results/pr-logs/pull/openshift_ovn-kubernetes/2314/pull-ci-openshift-ovn-kubernetes-master-e2e-gcp-ovn-techpreview/1849904878739001344

Therefore add a startup probe to give it a few seconds to fully start,
then run liveness probe. The test doesn't check for at least 8 seconds
(claimed in the comments) that the container never restarted, so this
should be fine to finish start up probe and get a few iterations of
liveness probe in.

Signed-off-by: Tim Rozet <trozet@redhat.com>