ipam: Fix init flow in case there are sticky ips in the system
In case the ovn control plane pod is restarted,
(for example, node restart, upgrade, or any pod restart reason),
the ipam claim Sync function created an array of all the ipams
belonging to a specific network, and used named allocator AllocateIPs
method to allocate them, so the allocator will reflect the current
cluster state.

The problem is AllocateIPs allows to allocate only one IP per given subnet,
so once it is used for all the IPs in the network at once, it will fail,
causing the control plane to have a crash loop.

Fix it by calling AllocateIPs per each claim on its own.
The claims were already created correctly, so each claim on its
own is safe to call AllocateIPs.

Add unit test to show the relevant assertion.

Signed-off-by: Or Shoval <oshoval@redhat.com>