ip allocator: rename AllocateIPs method
The `AllocateIPs` method name is misleading; it receives a slice of IPs
as parameter, which leads the user to think that it can allocate the IPs
in the slice in a single subnet - i.e. for subnet 10.0.0.0/24, it could
allocate IPs 10.0.0.2 and 10.0.0.3.

That is not the case: that function allocates a **single** IP address in
each of the subnets the IP allocator for said network manages. For
instance, on a dual-stack network, it could handle the following
subnets: 10.0.0.0/24 and fd10::/64. The function could then the invoked
with ips=10.0.0.2,fd10::2 to allocate **one** IP in each of the two
subnets managed by the IP allocator.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>