Update nad controller type names
Hoping to bring improvement to the naming and interfacing of the nad
controller package to reduce confusion. No functional changes but,
obviously, broad code impact. On to the details...

Rename package network-attach-def-controller to networkmanager. The
controller struct itself has been unexported and instead two interfaces
are provided: Controller to start and stop, and NetworkManager to
provide (for now) read-only information about networks. A default
implementation is provided that assumes the default network is the only
ever existing network, to be used when multi-network capabilities are
disabled or for testing. Other unexported structs have also been renamed
accordingly as well as other types that followed the former naming scheme.

Rename package network-controller-manager to controllemanager,
understood as a package that manages other controllers that can be of any
nature: network specific, network aware or network unaware. The
interfaces that controller managers need to implement to interface with
network manager have also been renamed accordingly as well as other
types that followed the former naming scheme.

For reasons beyond my comprehension, it befell to me coming accross, and
fixing, a number of unrelated issues, like flaking unit tests, missing
nftables stubbing to be able to run the tests on the testing container,
etc... that given the broad impact of this commit, I decided to fix
here instead of bothering with new commits. While I was at it, fixed all
go vet warnings on impacted files because it bothers me to have my IDE
painted yellow. We should add that to our linter. Moving on...

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>