Slightly fix UDN local gateway masquerade rules (and unit tests)
One of UDN local gateway masquerade rules was using an IP/length (eg.
169.254.0.2/29) rather than subnet/length (169.254.0.0/29). This is
ambiguous and we shouldn't depend on iptables interpreting it the way
we wanted.

Also, the corresponding unit test ended up being sort of silly because
it wasn't overriding the masquerade subnet to be "UDN-sized", so the
rules it output failed to actually distinguish the UDN and non-UDN
parts of the masquerade subnet. Fix that too.

Signed-off-by: Dan Winship <danwinship@redhat.com>