Improve egress service repair unit test
It was adding multiple incorrect rules for the same endpoint and
validating that they all got removed, but that won't work with the
nftables port where it will use a map, making it impossible to have
multiple rules for the same endpoint. Add another egress service to
the mix, with a different endpoint, so there are enough unique
endpoints for all the test cases.

Signed-off-by: Dan Winship <danwinship@redhat.com>