Fix a bug in FakeIPTables and a resulting bug in egressservice.Controller
The FakeIPTables implementation of List() was accidentally mutating
its own internal data to prepend "-A CHAINNAME" to each rule it
listed, and egressservice.Controller.repairIPTables() (which is tested
by the unit tests but not the e2e tests) ended up only working right
in the presence of that bug. Fix both.

Signed-off-by: Dan Winship <danwinship@redhat.com>