Enhance EgressQoS CR as a generic QoS feature
The current EgressQoS is tied to pod's egress traffic and may be used only for
DSCP marking. so this enhances the CR to be more generic enough so that it can
be easily extended for applying QoS for both east west and north south traffic
and supporting traffic policing and shaping.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>
Signed-off-by: Flavio Fernandes <ffernandes@nvidia.com>
Signed-off-by: Xiaobin Qu <xqu@nvidia.com>