Scopes egress firewall test for correct ACLs
With the change to no longer depend on mgmt port mac address annotation
on the node, the egress firewall test would unexpectedly now correctly
sync the management port during the test. This would cause
addAllowACLFromNode to execute and add an ACL for network policy to the
node switch. That in turn would cause this egress firewall test to fail
because it would see there are more than 0 ACLs on the switch.

This commit fixes it by properly scoping the lookup in the egress
firewall test for ACLs only applicable to egress firewall.

Signed-off-by: Tim Rozet <trozet@redhat.com>