(C)UDN CRD: add IPAM section and IPAM.Mode field.
IPAM options are only available for Layer2 network for now.
IPAM.Lifecycle is only supported when IPAM.Mode is Enabled (previously
expressed with non-empty subnets).
IPAM.Mode=Disabled is only supported for secondary network.
Update CEL: require or omit subnets based on the IPAM.Mode instead of
network role.

The only controller change needed is the location of IPAMLifecycle.
Empty subnets means turn off IPAM in NAD config, and CEL ensures that
subnets will be empty if and only if IPAM.Mode is Disabled.

Fix "should fail to render NAD" test, as it used to always return error
because of the empty target namespace instead of expected one.

Signed-off-by: Nadia Pinaeva <n.m.pinaeva@gmail.com>